+++ src/pbfbench/abc/tool/app.py
@@ -9,6 +9,7 @@ import datetime
 import logging
 from abc import ABC, abstractmethod
 from collections.abc import Callable
+from enum import StrEnum
 from pathlib import Path
 from typing import Annotated, TypeVar, cast, final
 
@@ -17,16 +18,18 @@ import typer
 import pbfbench.abc.app as abc_app
 import pbfbench.experiment.checks as exp_checks
 import pbfbench.experiment.file_system as exp_fs
-import pbfbench.experiment.in_progress as exp_in_progress
 import pbfbench.experiment.managers as exp_managers
+import pbfbench.experiment.report as exp_report
 import pbfbench.experiment.resume as exp_resume
 import pbfbench.experiment.run as exp_run
 import pbfbench.samples.status as smp_status
 import pbfbench.slurm.config as slurm_cfg
 from pbfbench import root_logging
 
+from . import bash as abc_tool_bash
 from . import config as abc_tool_cfg
 from . import connector as abc_tool_connector
+from . import environments as abc_tool_env
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -38,6 +41,13 @@ def log_filename() -> Path:
         + datetime.datetime.now(tz=None).strftime("%Y-%m-%dT%H_%M_%S")  # noqa: DTZ005
         + ".log",
     )
+
+
+class RichHelp(StrEnum):
+    """Rich help categories."""
+
+    MAIN_CMD = "Main commands"
+    UTILS_CMD = "Utilities"
 
 
 def build_application(
@@ -56,12 +66,20 @@ def build_application(
     # Run app
     #
     run_app = Run(connector_type)
-    app.command(name=run_app.NAME, help=run_app.help())(run_app.main)
+    app.command(
+        name=run_app.NAME,
+        help=run_app.help(),
+        rich_help_panel=RichHelp.MAIN_CMD,
+    )(run_app.main)
     #
     # Resume app
     #
     resume_app = Resume(connector_type)
-    app.command(name=resume_app.NAME, help=resume_app.help())(resume_app.main)
+    app.command(
+        name=resume_app.NAME,
+        help=resume_app.help(),
+        rich_help_panel=RichHelp.MAIN_CMD,
+    )(resume_app.main)
     #
     # Config app
     #
@@ -74,8 +92,29 @@ def build_application(
         config_app = ConfigAppWithArguments(
             cast("type[abc_tool_connector.WithArguments]", connector_type),
         )
-    app.command(name=config_app.NAME, help=config_app.help())(config_app.main)
-    # TODO add check when ready
+    app.command(
+        name=config_app.NAME,
+        help=config_app.help(),
+        rich_help_panel=RichHelp.UTILS_CMD,
+    )(config_app.main)
+    #
+    # Draft tool environment script app
+    #
+    draft_env_sh_app = ToolEnvWrapper(connector_type)
+    app.command(
+        name=draft_env_sh_app.NAME,
+        help=draft_env_sh_app.help(),
+        rich_help_panel=RichHelp.UTILS_CMD,
+    )(draft_env_sh_app.main)
+    #
+    # Report app
+    #
+    report_app = Report(connector_type)
+    app.command(
+        name=report_app.NAME,
+        help=report_app.help(),
+        rich_help_panel=RichHelp.UTILS_CMD,
+    )(report_app.main)
     return app
 
 
@@ -230,10 +269,10 @@ class Run[
         exp_manager: exp_managers.WithOptions,
     ) -> None:
         """Exit with error if the experiment is already running."""
-        match running_exp := exp_checks.experiment_is_running(
+        match running_exp := exp_report.running_experiment(
             exp_manager.data_fs_manager(),
         ):
-            case exp_checks.RunningExperiment():
+            case exp_report.RunningExperiment():
                 _LOGGER.critical(
                     "The experiment is already in progress\n"
                     "* Experiment launched the: %s\n"
@@ -360,18 +399,18 @@ class Resume[
             self.connector_type().description(),
             exp_name,
         )
-        if not data_fs_manager.in_progress_yaml().exists():
+        running_experiment = exp_report.running_experiment(data_fs_manager)
+        if running_experiment is None:
             _LOGGER.critical(
-                "The experiment `%s` is not in progress for the data directory `%s`",
+                "The experiment `%s` is not in progress for tool `%s` and topic `%s`",
                 exp_name,
-                data_fs_manager.root_dir(),
+                self.connector_type().description().name(),
+                self.connector_type().description().topic().name(),
             )
             raise typer.Exit(1)
-        data_in_progress = exp_in_progress.InDataDirectory.from_yaml(
-            data_fs_manager.in_progress_yaml(),
-        )
+
         work_fs_manager = exp_fs.WorkManager(
-            data_in_progress.working_directory(),
+            running_experiment.in_progress_data().working_directory(),
             self.connector_type().description(),
             exp_name,
         )
@@ -386,14 +425,14 @@ class Resume[
                     data_fs_manager,
                     work_fs_manager,
                     connector,
-                ), data_in_progress.job_id()
+                ), running_experiment.in_progress_data().job_id()
             case abc_tool_connector.WithArguments():
                 return exp_managers.WithArguments(
                     exp_name,
                     data_fs_manager,
                     work_fs_manager,
                     connector,
-                ), data_in_progress.job_id()
+                ), running_experiment.in_progress_data().job_id()
             case exp_checks.RunErrors():
                 _LOGGER.critical("Cannot instantiate connector")
                 raise typer.Exit(1)
@@ -449,6 +488,7 @@ class ConfigAppWithOptions[
 
         config.to_yaml(config_exp_yaml)
         _LOGGER.info("Tool configuration written to %s", config_exp_yaml)
+        raise typer.Exit(0)
 
     @abstractmethod
     def _fake_config(self) -> Config:
@@ -495,3 +535,167 @@ class ConfigAppWithArguments(
                 "$input_experiment_name",
             )
         return abc_tool_cfg.Arguments(args)
+
+
+class ToolEnvWrapperOpts:
+    """Tool environnment wrapper options."""
+
+    ERASE = typer.Option(
+        "--erase/--do-not-erase",
+        help="Erase existing tool environnment wrapper script",
+    )
+
+
+class ToolEnvWrapper[CType: type[abc_tool_connector.WithOptions]]:
+    """Draft tool environnment wrapper script."""
+
+    NAME = "draft-env"
+
+    def __init__(self, tool_connector_type: CType) -> None:
+        self._tool_connector_type = tool_connector_type
+
+    def main(
+        self,
+        data_dir: Annotated[Path, Arguments.DATA_DIR],
+        erase: Annotated[bool, ToolEnvWrapperOpts.ERASE] = False,
+        debug: Annotated[bool, root_logging.OPT_DEBUG] = False,
+    ) -> None:
+        """Generate environnment wrapper script."""
+        root_logging.init_logger(
+            _LOGGER,
+            (
+                f"Generating an environnment wrapper script"
+                f" for tool {self._tool_connector_type.description().name()}"
+                f" for topic {self._tool_connector_type.description().topic().name()}"
+            ),
+            debug,
+        )
+        data_exp_manager = exp_fs.DataManager(
+            data_dir.resolve(),
+            self._tool_connector_type.description(),
+            "fake_exp",
+        )
+        core_command_sh_path = abc_tool_bash.CommandsWithOptions.core_command_sh_path(
+            self._tool_connector_type.description(),
+        )
+        env_wrapper_sh_path = data_exp_manager.tool_env_script_sh()
+        if env_wrapper_sh_path.exists() and not erase:
+            _LOGGER.critical(
+                "The environnment wrapper script %s already exists",
+                env_wrapper_sh_path,
+            )
+            raise typer.Exit(1)
+
+        if not data_exp_manager.tool_dir().exists():
+            _LOGGER.info(
+                "Create non-existing tool directory %s",
+                data_exp_manager.tool_dir(),
+            )
+            data_exp_manager.tool_dir().mkdir(parents=True)
+
+        to_do_lines: list[str] = []
+        with core_command_sh_path.open() as f_in:
+            in_block_comment = True
+            iter_lines = iter(f_in)
+            while in_block_comment:
+                line = next(iter_lines, None)
+                if line is None or not line.startswith("#"):
+                    in_block_comment = False
+                else:
+                    to_do_lines.append(line.rstrip())
+
+        with env_wrapper_sh_path.open("w") as f_out:
+            f_out.write("# Tips from the tool core commands:\n")
+            f_out.write("\n".join(to_do_lines))
+            if to_do_lines:
+                f_out.write("\n")
+            f_out.write("\n")
+            f_out.write(abc_tool_env.BashEnvWrapper.BEGIN_ENV_MAGIC_COMMENT + " ===\n")
+            f_out.write("\n")
+            f_out.write("# Commands to initialize the environment\n")
+            f_out.write("\n")
+            f_out.write(abc_tool_env.BashEnvWrapper.MID_ENV_MAGIC_COMMENT + " ---\n")
+            f_out.write("\n")
+            f_out.write("# Commands to close the environment\n")
+            f_out.write("\n")
+            f_out.write(abc_tool_env.BashEnvWrapper.END_ENV_MAGIC_COMMENT + " ===\n")
+
+        _LOGGER.info(
+            "Tool environnment wrapper draft script written to %s",
+            env_wrapper_sh_path,
+        )
+        raise typer.Exit(0)
+
+    def help(self) -> str:
+        """Get help string."""
+        return (
+            "Generate a draft environnment wrapper script"
+            f" for tool {self._tool_connector_type.description().name()}"
+            f" for topic {self._tool_connector_type.description().topic().name()}"
+        )
+
+
+class ReportOpts:
+    """Report command options."""
+
+    YAML_FILE = typer.Option(
+        "--yaml",
+        "-o",
+        help="YAML output file",
+    )
+
+
+class Report[CType: type[abc_tool_connector.WithOptions]]:
+    """Report command."""
+
+    NAME = "report"
+
+    def __init__(self, tool_connector_type: CType) -> None:
+        self._tool_connector_type = tool_connector_type
+
+    def main(
+        self,
+        exp_name: Annotated[str, Arguments.EXP_NAME],
+        data_dir: Annotated[Path, Arguments.DATA_DIR],
+        yaml_file: Annotated[Path | None, ReportOpts.YAML_FILE] = None,
+        debug: Annotated[bool, root_logging.OPT_DEBUG] = False,
+    ) -> None:
+        """Generate report."""
+        root_logging.init_logger(
+            _LOGGER,
+            (
+                f"Generating a report"
+                f" for tool {self._tool_connector_type.description().name()}"
+                f" for topic {self._tool_connector_type.description().topic().name()}"
+            ),
+            debug,
+        )
+        data_exp_manager = exp_fs.DataManager(
+            data_dir.resolve(),
+            self._tool_connector_type.description(),
+            exp_name,
+        )
+        if not data_exp_manager.exp_dir().exists():
+            _LOGGER.critical(
+                "There is no experiment nammed `%s` for tool `%s` and topic `%s`",
+                data_exp_manager.experiment_name(),
+                self._tool_connector_type.description().name(),
+                self._tool_connector_type.description().topic().name(),
+            )
+            raise typer.Exit(1)
+
+        report = exp_report.Report.from_data_exp_fs_manager(data_exp_manager)
+
+        _LOGGER.info("Report:\n%s", exp_report.report_to_string(report))
+        if yaml_file is not None:
+            _LOGGER.info("Writing report to `%s`", yaml_file)
+            report.to_yaml(yaml_file)
+        raise typer.Exit(0)
+
+    def help(self) -> str:
+        """Get help string."""
+        return (
+            "Generate a report"
+            f" for tool {self._tool_connector_type.description().name()}"
+            f" for topic {self._tool_connector_type.description().topic().name()}"
+        )